requirements.txt
asgiref==3.7.2
cffi==1.15.1
distlib==0.3.7
Django==3.2.6
django-geojson==4.0.0
django-jazzmin==2.6.0
django-leaflet==0.29.0
django-tinymce==3.6.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.4
geojson==3.0.1
gunicorn==21.2.0
h3==3.7.6
leaflet==0.0.3
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
platformdirs==3.10.0
psycopg2==2.9.7
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
timezonefinder==6.2.0
typing_extensions==4.9.0
tzdata==2023.3
virtualenv==20.24.5
whitenoise==6.5.0
xlrd==2.0.1
